itec_app/requirements.txt
APScheduler==3.10.1
asgiref==3.8.1
Django==5.0.3
djangorestframework==3.15.1
et-xmlfile==1.1.0
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
Pillow==9.5.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.0.1
